move schema sampling into vPartition
